backend/pyproject.toml
[project]
name = "thinkaloud-backend"
version = "2.0.0"
description = "ThinkAloud.ai — AI-powered interview practice platform"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # FastAPI & Server
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "python-multipart>=0.0.18",

    # Database
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "psycopg2-binary>=2.9.0",

    # Redis
    "redis>=5.0.0",

    #trace 
    "ddtrace>=2.0.0",
    "structlog>=24.0.0",

    #docker
    "docker>=7.0.0",



    # Authentication
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]==1.7.4",
    "bcrypt==4.1.2",

    # Validation & Settings
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "email-validator>=2.0.0",

    # AI & Voice
    "httpx>=0.28.0",
    "livekit-api>=1.0.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
    "ruff>=0.9.0",
]
